correctly showing badge
