+++ tests/e2e/fixtures/enabled-test/disabled/index.html
@@ -39,10 +39,9 @@
   
   <p>This is the root page of the "disabled" section.</p>
   <p>Since this page is not under the path where the feedback package is created, it should remain disabled for BackChannel.</p>
+
   
-  <script>
-    // This script is just a placeholder for the test to verify page loading
-    console.log('Disabled root page loaded');
-  </script>
+  <!-- Load BackChannel -->
+  <script type="module" src="/src/index.ts"></script>
 </body>
 </html>
