+++ tests/e2e/fixtures/enabled-test/disabled/subdir/index.html
@@ -40,9 +40,7 @@
   <p>This is a subdirectory page under the "disabled" section.</p>
   <p>Since this page is not under the path where the feedback package is created, it should remain disabled for BackChannel.</p>
   
-  <script>
-    // This script is just a placeholder for the test to verify page loading
-    console.log('Disabled subdirectory page loaded');
-  </script>
+  <!-- Load BackChannel -->
+  <script type="module" src="/src/index.ts"></script>
 </body>
 </html>
