+++ tests/e2e/fixtures/enabled-test/enabled/index.html
@@ -40,9 +40,7 @@
   <p>This is the root page of the "enabled" section.</p>
   <p>The test will create a feedback package at this location, which should make all pages under this path enabled for BackChannel.</p>
   
-  <script>
-    // This script is just a placeholder for the test to verify page loading
-    console.log('Enabled root page loaded');
-  </script>
+  <!-- Load BackChannel -->
+  <script type="module" src="/src/index.ts"></script>
 </body>
 </html>
