+++ tests/e2e/fixtures/enabled-test/enabled/subdir/index.html
@@ -40,9 +40,7 @@
   <p>This is a subdirectory page under the "enabled" section.</p>
   <p>Since this page is under the path where the feedback package is created, it should be enabled for BackChannel.</p>
   
-  <script>
-    // This script is just a placeholder for the test to verify page loading
-    console.log('Enabled subdirectory page loaded');
-  </script>
+  <!-- Load BackChannel -->
+  <script type="module" src="/src/index.ts"></script>
 </body>
 </html>
