+++ tests/e2e/fixtures/enabled-test/index.html
@@ -39,9 +39,7 @@
   <p>This is the root page for testing BackChannel enabled/disabled state detection.</p>
   <p>The test will create a feedback package at the root of the "enabled" section, which should make all pages under that path enabled for BackChannel.</p>
   
-  <script>
-    // This script is just a placeholder for the test to verify page loading
-    console.log('Root page loaded');
-  </script>
+  <!-- Load BackChannel -->
+  <script type="module" src="/src/index.ts"></script>
 </body>
 </html>
